Add auto-generation files
Add create-day script and template.rs. I've been using these for about a
week now, and figured they would be good to have in the repo from now on
so that I have a place to track them for later years and in case someone
else wants to use this repo for making their own Rust solutions (for
some reason).

Signed-off-by: Ben Cheatham <benhartcheatham@gmail.com>