Day twenty part 2 solution
Technically first try, but I used someone else's code to check both
parts today.

My intuition that you needed to compute cycles was right, and even that
on the part(s) above "rx" mattered, but I still read a guide for this
part. Took a bit of finagling, and the solution(s) are pretty icky to
look at (compared to the comparison code I was looking at :/), but it
works.

Signed-off-by: Ben Cheatham <Benjamin.Cheatham@amd.com>