Day twenty five part 1 solution
First try, technically.

I looked up a solution because min-cut was a pain to try to implement,
this is also after one rewrite. This solution is quite slow, I imagine
due to having to iterate through the edges every time I want to find an
edge for a node, but it could still be sped up with some refactoring.
Who cares though, it's the final day!

Signed-off-by: Ben Cheatham <Benjamin.Cheatham@amd.com>